Апка/services/tasks/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.14.0
pydantic-settings==2.6.1
redis==5.0.8
aio-pika==9.4.3
python-jose[cryptography]==3.3.0
httpx==0.27.2
email-validator==2.1.1
